Update ActivityLogForm.tsx
FAB log entry formStep 2: Update ActivityLogForm.tsx
We need to make a small change to your existing form to hide the parts we don't need in the pop-up (like the list of today's activities). We'll add an isModal prop to control this.

Action: Replace the code in src/components/ActivityLogForm.tsx with this updated version.